Create ecdf module
Create "ecdf" module in "visualize" subpackage.
remove previous functions
