--- plantcv/plantcv/visualize/eCDF/__init__.py
@@ -1,5 +1,0 @@
-from plantcv.plantcv.visualize.ecdf.obj_size import obj_size
-from plantcv.plantcv.visualize.ecdf.pix_intensity import pix_intensity
-
-
-__all__ = ["obj_size","pix_intensity"]
